switch to single suite for stop control
